day-2/day-2a/requirements.txt
# Google ADK - Agent Development Kit
google-adk>=0.1.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Additional utilities
# requests>=2.31.0  # For API calls
# pandas>=2.0.0     # For data processing
